Feat: Remove language configuration files ready for using vscode's internal configs.
Also, change package.json:

- Version bump.
- Fix user settings descriptions.
- Remove language contribution declaration.
- Added 3 new user settings:
`multiLineStyleBlocks` to add unsupported languages to use the `/** */` comments;

`overrideDefaultLanguageMultiLineComments` to override the style of multi-line comments for a language for the native vscode `blockComment` command.

`bladeOverrideComments` to switch between HTML style `<!-- -->` comments, and Blade style `{{--  --}}` comments, when the file is using the Blade language. Also added a new keybinding for this.

- Updated various dependencies and added new ones.