fix: certain langConfig property values from preventing activating the extension, and more.
Fix:

- The `for...of` loops in `configureCommentBlocks` method to use the `singleLineLangs` variable which uses the `getSingleLineLanguages` method instead of directly using the property `singleLineBlocksMap`.

- The single line languages `for...of` loop to only setup the config for the langId if it's not currently set as disabled in the user config settings, using the `isLangIdDisabled` method to check.

- The merging of onEnterRules for single-line comments. When a language config doesn't have the `onEnterRules` key and the user config setting `singleLineBlockOnEnter` is true, previously the extension couldn't be activated due to the error:

> Cannot read properties of undefined (reading 'concat')

because the key is undefined. Fixed by using the chaining operator (?.) before trying to access the array method `concat`. If the key exists then concat still works and combines the new rules with the internal config rules. If key is undefined then by using the nullish coalescing operator (??) we just assign the new rules since there's no rules to combine with.

- When a config has the `wordPattern` key with a regex pattern as a string,
the extension can't activate due to the error:

> TypeError: e.exec is not a function

It's unknown why the allowed regex as string causes this, there seems to be a similar, related issue in https://github.com/microsoft/vscode/issues/171194 but was closed as (wrongly?) an invalid issue.

Fixed by changing the string to an actual regex pattern. But because `wordPattern` can also be an object with a key of `pattern` then we also need to change this and add the regex pattern directly as the value of `wordPattern` instead of an object.

- The extension only ran on opening a document, but we need it to run everytime the current editor changes or the current language changes. Fixed to use the on change event `onDidChangeTextDocument` instead.

Removed:

- Remove the unnecessary check for the length of landId is more than 0 in `setMultiLineCommentLanguageDefinitions` method, because just the first part of that conditional also checks for not NULL or empty strings, so the extra check is redundant.

Added:

- The return type of `getExtensionNames` method, now as an object with `id`, `name`, and `displayName` keys, so TS isn't guessing at compile time.

- Code comments

- The execution of changing the blockComment style when it's been overridden for that langId in the user config settings.