fix: revert the `onDidChangeTextDocument` back to `onDidOpenTextDocument`
In commit 30de00a, the `onDidOpenTextDocument` workspace event in `extension.ts` was changed to `onDidChangeTextDocument`, with the assumption that the former only ran the function when opening a document. However, this is not the case. It runs when a document is opened OR when the language ID has been changed. It's fully described in https://github.com/microsoft/vscode/blob/4e8fbaef741afebd24684b88cac47c2f44dfb8eb/src/vscode-dts/vscode.d.ts#L13716-L13728

This also fixes the issue of running the extension on every document change, even if it's not a comment, which would eventually cause performance issues.