fix: `//` onEnterRule regex to stop it preventing the `//!` from matching it's regex.
When using `//!`-style and the `singleLineBlockOnEnter` setting is turned on, the `//` regex matches because the negative lookahead `(?!...)` expression doesn't include the `!` therefore it would still match the `//` part of the `//!`. Therefore, it would enter `//` on the next line instead of the anticipated `//!`-style.

Fixed by updating the regex to include the `!` in the negative lookahead expression which basically means "don't match if it has these characters in".