fix: single-line onEnterRules were not working for any languages when `singleLineBlockOnEnter` was set to `true`.
For some unknown reason, when the language config had regexes and the config is set via
`vscode.languages.setLanguageConfiguration()`, the regexes weren't actually been used. By reconstructing the regex before the config is set, it fixes the issue. This also helps when the internal config has some malformed regexes, or regex as strings.

Also, in JavaScript and TypeScript files, pressing tab on an empty line may insert a `*` on that line. Probably the same bug as above, which fixes it. Fixes Tab writes * every time I press yCodeTech/auto-comment-blocks#2