fix: single-line comments `onEnterRules` not adding a new commented line when cursor is in the middle of the text.
When `singleLineBlockOnEnter` setting is `true`, and when then cursor is in the middle of the text, eg. `// text | text`, the comment would not be continued on the next line when enter is pressed. Therefore the text after the cursor would not be a comment on the new line and be classed as code. This makes commenting tedious when keeping it clean and tidy with max line lengths.

This is fixed by adding new regex rules to identify when there is text after the cursor. The rules are now merged and duplicates removed instead of just `concat`ing.

Changed the name of `mergeConfig` method to `mergeConfigProperty` to be more appropriate naming, and refactored it to be DRYer and be able to merge any of the config's property arrays. Changed the method references accordingly.