fix: `Logger::logMessage` not logging regex meta data.
When a log's extra `debug` data contains a `RegExp` object, `JSON.stringify` can not convert it to a string, so it just outputs an empty object `{}` in the Output channel. This is useless for debugging.

Fixed by adding a new `Logger::metaReplacer` method to use as a replacer function in the `JSON.stringify` method's replacer arg.
This new method will convert all instances of `RegExp` to it's string representation ready for output in the channel.