style: add empty new lines inbetween the language config filepaths.
Because the language config filepaths are long and spill over 2 lines, it's quite hard to distinguish the lines from one another.

So to fix for style purposes, find the correct meta data by checking the `message` for the string `"filepaths"` as in `"The language config filepaths found are:"`; then split the data by comma (the comma separating each item). Then join the resulting array back with the comma and adding an empty new line too (`\n`).

This makes the filepaths so much easier to read.