feat: add the ability to get all Windows extensions when running on WSL.
This is a work around for the fact that the `vscode.extensions.all` API can't find any built-in extensions on the Windows-side when running in WSL. It only gets the WSL-installed extensions, which caused the extension not to work in yCodeTech/auto-comment-blocks#6

There is an open issue and a proposed API change in the sourcecode since 2022, but hasn't been released as a public API. See https://github.com/microsoft/vscode/issues/145307 for more info.

Added:

- Added new `getExtensionsPathsFromWslEnv` method to retrieve the Windows built-in and user extensions paths from the environment variables when running in WSL.

- Added new `readExtensionsFromDirectory` method to read the extensions paths and retrieve each extension data.

Changed:

- Changed `findAllLanguageConfigFilePaths` method to check if the extension is running in WSL. If it `isWsl`, then it calls the new `getExtensionsPathsFromWslEnv` method to get the paths, and uses the new `readExtensionsFromDirectory` to obtain the extensions data.

We construct a new `extensions` array with all the windows-installed built-in and user extensions and also the WSL-installed extensions via `vscode.extensions.all`. (The extension will still work on Windows without WSL due to it using `vscode.extensions.all` outside of the `isWsl` check.)

- Changed the `for...of` loop to iterate through the new `extensions` array.