fix: `logDebugInfo` user extensions path to use `getExtensionData`.
Previously we manually created the user extensions path from the home directory env variables and concatenated strings on to it. This is only good if the default extensions directory is used. This will fail if the user has changed the directory location with the CLI `code --extensions-dir <path>`

So to fix, we're now getting the user extensions path from where ever this extension is located. The code for this is defined in `setExtensionData` method as committed in 25fab48.

- Changed the construction of the users extensions path in `logDebugInfo` to use the `getExtensionData` method to get the path.

- Removed the unneeded home directory variables.