fix: "includes is not a function" error when `lineComment` is an object.
In the `setSingleLineCommentLanguageDefinitions` method, it uses the `String.includes` method to detect when the `lineComment` string has `includes` a semi-colon - ;. Apparently the lineComment key in the language config can be either a string or an object. The object has a `comment` key, among others. But because the `includes` method is on the String prototype, when its used on the object, it errors out with "lineComment.includes is not a function". The only language to do this was "makefile".

To fix:

- Added a check to see if the `lineComment` is an object with a `comment` key inside, if it does then use the string value of the `comment` key. Otherwise, it will use the string value of the `lineComment`.