fix: [BUG]: single-line comment fails in VS Code June release Fixes yCodeTech/auto-comment-blocks#10
In the `setSingleLineCommentLanguageDefinitions` method, it uses the `String.includes` method to detect when the `lineComment` string `includes` a semi-colon - ;.  The `lineComment` the language config has changed to be either a string or an object. The object has a `comment` and `noIndent` keys.

The `includes` method fails due to it being an object and errors out with "lineComment.includes is not a function".

To fix:

- Added a check to see if the `lineComment` is an object with a `comment` key inside, if it does then use the string value of the `comment` key. Otherwise, it will use the string value of the `lineComment`.