ci(setup-labels): fix running on every push.
This workflow only needs to run once (ideally) to setup the labels. Currently it runs on every push.

Fixed by changing it to run only on this specific file `path`, which will run when first committing the file, and on subsequent modifications of the file.
Also added a manual way via `workflow_dispatch` event, so clicking the "Run workflow" button in the GitHub Actions tab will run the workflow.