fix: TS compile error: Cannot find name 'id'.
https://github.com/yCodeTech/auto-comment-blocks/actions/runs/16339281667/job/46157703386?pr=12

This error occurs because testing used a hard-coded path to simulate what the `vscode.extensions.getExtension().extensionPath` API would get. Since commit b0a40aa changed the way we get/set the package.json data, this was just an oversight.

Since we're already using the path of the extension to get the package.json in `getExtensionPackageJsonData`, we can just leverage this, and properly integrate the path.

Fixed by:
- Adding the extension path as a new `extensionPath` key in the `packageJSON` data object.

- Completely removing the VS Code's `.extensions.getExtension()` API call in `createExtensionData` method and changing it to use the `extensionPath` key in the `packageJsonData` property.