fix: `languageConfigFilePaths` to support multiple config paths per lang
Due to some extensions either not specifying all the language configuration properties in the file or setting the comment properties to an empty array (like the "stillat-llc.vscode-antlers" extension does which overrides the HTML config and makes the comments not work without the extension activated), we need to ensure that they are merged with other possible configs for the language.

- Changed `languageConfigFilePaths` Map to accept an array value instead of just a single string.

- Changed `findAllLanguageConfigFilePaths` method to set the config path into an array as the value in `languageConfigFilePaths` Map, and to push a new path into the array if the `langId` key already exists in the Map.

- Changed `setLanguageConfigDefinitions` method to allow it to loop through the new array of paths which sets the configs into the `languageConfigs` Map, and if the `langId` key already exists then merge the existing config with the new config.

This merges the comments from both configs together and skips empty arrays. It also merges all other config properties together to ensure we are getting a full language configuration. By merging all available configs for the language, the language won't have any issues regarding malformed configs, or comments not working.