feat: show a user notification for JSON parsing failures
Because `readJsonFile` is a make or break kinda function, we should notify the user by showing an error popup dialog.

- Added vscode's `showErrorMessage` method to tell the user that errors occurred when parsing JSON files, and the extension won't work correctly. It also opens the Output Channel when the button is clicked.

- Enabled the error to be thrown. By `throw`ing the error, we allow the extension to fail startup and crash at the point of parse errors.

This is to prevent further code that requires the parsed json from the `readJsonFile` function from throwing an unrelated error such as the "Cannot convert undefined or null to object" error, which makes it harder to debug. So if we fail the extension here, then we can easily debug the root cause.